Fixed test issue, beautified code
The issue was the hashPassword stub did not generated any string for `undefined` salt. When this was fixed, the tests passed.

For every new created route (thanks to Illya) the prettier was run and the code looks more shiny. Also I got rid of forEach, replaced it with `for of` construction